+++ folded/ConstructorReferenceNotationWithConstTestData-folded.java
@@ -1,0 +1,98 @@
+package data;
+
+import java.util.*;
+
+public class ConstructorReferenceNotationWithConstTestData {
+
+    static class Const {
+        const ConstClass SELF = ::new;
+        const ConstClass SELF_ANN = ::new{};
+        const ConstClass SELF_SUB = new SubConstClass();
+        const ConstClass SELF_SUB_ANN = new SubConstClass() {
+        };
+
+        private const HashMap<String, String> MAP = ::new;
+        private const HashMap<String, String> MAP2 = ::new;
+        private const Map<String, String> MAP3 = new HashMap<>();
+        private const Map<String, String> MAP_TREE = new TreeMap<>();
+        private const Map<String, String> MAP4 = Map.of();
+
+        private const List<String> LIST = new ArrayList<>();
+        private const List<String> LIST2 = List.of();
+        private const List<String> LIST_SINGLE = List.of("1");
+        private const List<String> LIST_LINKED = new LinkedList<>();
+
+
+        const ConstClass SELF_PARAM_1 = ::new(true);
+        const ConstClass SELF_PARAM_2 = ::new(false, LIST_SINGLE.get(0));
+
+        const ConstClass SELF_SUBCLASS_MORE_FIELD = new ConstClass() {
+            int i = 1;
+        };
+        const ConstClass SELF_SUBCLASS_MORE_FUNC = new ConstClass() {
+            public void setOk(boolean ok) {
+            }
+        };
+
+        protected static ConstClass SELF_NULL = null;
+        protected static ConstClass EMPTY;
+    }
+    static class Fields {
+        final ConstClass SELF = ::new;
+        ConstClass SELF_ANN = ::new{};
+        public final ConstClass SELF_SUB = new SubConstClass();
+        public final ConstClass SELF_SUB_ANN = new SubConstClass() {
+        };
+
+        private final HashMap<String, String> MAP = ::new;
+        private final HashMap<String, String> MAP2 = ::new;
+        private final Map<String, String> MAP3 = new HashMap<>();
+
+        private final List<String> LIST = new ArrayList<>();
+        private final List<String> LIST2 = List.of("1");
+
+        public final ConstClass SELF_PARAM_1 = ::new(true);
+        public final ConstClass SELF_PARAM_2 = ::new(false, LIST2.get(0));
+
+        public final ConstClass SELF_SUBCLASS_MORE_FIELD = new ConstClass() {
+            int i = 1;
+        };
+        public final ConstClass SELF_SUBCLASS_MORE_FUNC = new ConstClass() {
+            public void setOk(boolean ok) {
+            }
+        };
+
+        protected ConstClass SELF_NULL = null;
+        protected ConstClass EMPTY;
+    }
+
+    static class ConstClass {
+        boolean ok;
+        String string;
+
+        public ConstClass() {
+        }
+
+        public ConstClass(boolean ok) {
+            this.ok = ok;
+        }
+
+        public ConstClass(boolean ok, String string) {
+            this.ok = ok;
+            this.string = string;
+        }
+
+        @Override
+        public String toString() {
+            return new StringJoiner(", ", ConstClass.class.getSimpleName() + "[", "]")
+                    .add("string='" + string + "'")
+                    .toString();
+        }
+    }
+
+    static class SubConstClass extends ConstClass {
+    }
+
+    static final class SubConstClass2 extends ConstClass {
+    }
+}