+++ testData/ConstructorReferenceNotationTestData.java
@@ -1,0 +1,100 @@
+package data;
+
+import java.util.*;
+
+public class ConstructorReferenceNotationTestData {
+
+    static class Const <fold text='{...}' expand='true'>{
+        public static final ConstClass SELF = new ConstClass();
+        public static final ConstClass SELF_ANN = new ConstClass() <fold text='{...}' expand='true'>{
+        }</fold>;
+        public static final ConstClass SELF_SUB = new SubConstClass();
+        public static final ConstClass SELF_SUB_ANN = new SubConstClass() <fold text='{...}' expand='true'>{
+        }</fold>;
+
+        private static final HashMap<String, String> MAP = new HashMap<>();
+        private static final HashMap<String, String> MAP2 = new HashMap<fold text='<~>' expand='false'><String, String></fold>();
+        private static final Map<String, String> MAP3 = new HashMap<>();
+        private static final Map<String, String> MAP_TREE = new TreeMap<>();
+        private static final Map<String, String> MAP4 = Map.of();
+
+        private static final List<String> LIST = new ArrayList<>();
+        private static final List<String> LIST2 = List.of();
+        private static final List<String> LIST_SINGLE = List.of("1");
+        private static final List<String> LIST_LINKED = new LinkedList<>();
+
+
+        public static final ConstClass SELF_PARAM_1 = new ConstClass(true);
+        public static final ConstClass SELF_PARAM_2 = new ConstClass(false, LIST_SINGLE.get(0));
+
+        public static final ConstClass SELF_SUBCLASS_MORE_FIELD = new ConstClass() <fold text='{...}' expand='true'>{
+            int i = 1;
+        }</fold>;
+        public static final ConstClass SELF_SUBCLASS_MORE_FUNC = new ConstClass() <fold text='{...}' expand='true'>{
+            public void setOk(boolean ok) <fold text='{}' expand='true'>{
+            }</fold>
+        }</fold>;
+
+        protected static ConstClass SELF_NULL = null;
+        protected static ConstClass EMPTY;
+    }</fold>
+    static class Fields <fold text='{...}' expand='true'>{
+        final ConstClass SELF = new ConstClass();
+        ConstClass SELF_ANN = new ConstClass() <fold text='{...}' expand='true'>{
+        }</fold>;
+        public final ConstClass SELF_SUB = new SubConstClass();
+        public final ConstClass SELF_SUB_ANN = new SubConstClass() <fold text='{...}' expand='true'>{
+        }</fold>;
+
+        private final HashMap<String, String> MAP = new HashMap<>();
+        private final HashMap<String, String> MAP2 = new HashMap<fold text='<~>' expand='false'><String, String></fold>();
+        private final Map<String, String> MAP3 = new HashMap<>();
+
+        private final List<String> LIST = new ArrayList<>();
+        private final List<String> LIST2 = List.of("1");
+
+        public final ConstClass SELF_PARAM_1 = new ConstClass(true);
+        public final ConstClass SELF_PARAM_2 = new ConstClass(false, LIST2.get(0));
+
+        public final ConstClass SELF_SUBCLASS_MORE_FIELD = new ConstClass() <fold text='{...}' expand='true'>{
+            int i = 1;
+        }</fold>;
+        public final ConstClass SELF_SUBCLASS_MORE_FUNC = new ConstClass() <fold text='{...}' expand='true'>{
+            public void setOk(boolean ok) <fold text='{}' expand='true'>{
+            }</fold>
+        }</fold>;
+
+        protected ConstClass SELF_NULL = null;
+        protected ConstClass EMPTY;
+    }</fold>
+
+    static class ConstClass <fold text='{...}' expand='true'>{
+        boolean ok;
+        String string;
+
+        public ConstClass() <fold text='{}' expand='true'>{
+        }</fold>
+
+        public ConstClass(boolean ok)<fold text=' { ' expand='false'> {
+            </fold>this.ok = ok;<fold text=' }' expand='false'>
+        }</fold>
+
+        public ConstClass(boolean ok, String string) <fold text='{...}' expand='true'>{
+            this.ok = ok;
+            this.string = string;
+        }</fold>
+
+        @Override
+        public String toString() <fold text='{...}' expand='true'>{
+            return new StringJoiner(", ", ConstClass.class.getSimpleName() + "[", "]")
+                    .add("string='" + string + "'")
+                    .toString();
+        }</fold>
+    }</fold>
+
+    static class SubConstClass extends ConstClass <fold text='{...}' expand='true'>{
+    }</fold>
+
+    static final class SubConstClass2 extends ConstClass <fold text='{...}' expand='true'>{
+    }</fold>
+}