+++ testData/ConstructorReferenceNotationWithConstTestData.java
@@ -1,0 +1,100 @@
+package data;
+
+import java.util.*;
+
+public class ConstructorReferenceNotationWithConstTestData {
+
+    static class Const <fold text='{...}' expand='true'>{
+        <fold text='const' expand='false'>public static final</fold> ConstClass SELF =<fold text=' ::' expand='true'> </fold>new<fold text='' expand='true'> </fold><fold text='' expand='true'>ConstClass</fold><fold text='' expand='true'>()</fold>;
+        <fold text='const' expand='false'>public static final</fold> ConstClass SELF_ANN =<fold text=' ::' expand='true'> </fold>new<fold text='' expand='true'> </fold><fold text='{}' expand='true'>ConstClass<fold text='' expand='true'>()</fold> <fold text='{...}' expand='true'>{
+        }</fold></fold>;
+        <fold text='const' expand='false'>public static final</fold> ConstClass SELF_SUB = new SubConstClass();
+        <fold text='const' expand='false'>public static final</fold> ConstClass SELF_SUB_ANN = new SubConstClass() <fold text='{...}' expand='true'>{
+        }</fold>;
+
+        private <fold text='const' expand='false'>static final</fold> HashMap<String, String> MAP =<fold text=' ::' expand='true'> </fold>new<fold text='' expand='true'> </fold><fold text='' expand='true'>HashMap<></fold><fold text='' expand='true'>()</fold>;
+        private <fold text='const' expand='false'>static final</fold> HashMap<String, String> MAP2 =<fold text=' ::' expand='true'> </fold>new<fold text='' expand='true'> </fold><fold text='' expand='true'>HashMap<fold text='<~>' expand='false'><String, String></fold><fold text='' expand='true'></fold>()</fold>;
+        private <fold text='const' expand='false'>static final</fold> Map<String, String> MAP3 = new HashMap<>();
+        private <fold text='const' expand='false'>static final</fold> Map<String, String> MAP_TREE = new TreeMap<>();
+        private <fold text='const' expand='false'>static final</fold> Map<String, String> MAP4 = Map.of();
+
+        private <fold text='const' expand='false'>static final</fold> List<String> LIST = new ArrayList<>();
+        private <fold text='const' expand='false'>static final</fold> List<String> LIST2 = List.of();
+        private <fold text='const' expand='false'>static final</fold> List<String> LIST_SINGLE = List.of("1");
+        private <fold text='const' expand='false'>static final</fold> List<String> LIST_LINKED = new LinkedList<>();
+
+
+        <fold text='const' expand='false'>public static final</fold> ConstClass SELF_PARAM_1 =<fold text=' ::' expand='true'> </fold>new<fold text='' expand='true'> </fold><fold text='' expand='true'>ConstClass</fold>(true);
+        <fold text='const' expand='false'>public static final</fold> ConstClass SELF_PARAM_2 =<fold text=' ::' expand='true'> </fold>new<fold text='' expand='true'> </fold><fold text='' expand='true'>ConstClass</fold>(false, LIST_SINGLE.get(0));
+
+        <fold text='const' expand='false'>public static final</fold> ConstClass SELF_SUBCLASS_MORE_FIELD = new ConstClass() <fold text='{...}' expand='true'>{
+            int i = 1;
+        }</fold>;
+        <fold text='const' expand='false'>public static final</fold> ConstClass SELF_SUBCLASS_MORE_FUNC = new ConstClass() <fold text='{...}' expand='true'>{
+            public void setOk(boolean ok) <fold text='{}' expand='true'>{
+            }</fold>
+        }</fold>;
+
+        protected static ConstClass SELF_NULL = null;
+        protected static ConstClass EMPTY;
+    }</fold>
+    static class Fields <fold text='{...}' expand='true'>{
+        final ConstClass SELF =<fold text=' ::' expand='true'> </fold>new<fold text='' expand='true'> </fold><fold text='' expand='true'>ConstClass</fold><fold text='' expand='true'>()</fold>;
+        ConstClass SELF_ANN =<fold text=' ::' expand='true'> </fold>new<fold text='' expand='true'> </fold><fold text='{}' expand='true'>ConstClass<fold text='' expand='true'>()</fold> <fold text='{...}' expand='true'>{
+        }</fold></fold>;
+        public final ConstClass SELF_SUB = new SubConstClass();
+        public final ConstClass SELF_SUB_ANN = new SubConstClass() <fold text='{...}' expand='true'>{
+        }</fold>;
+
+        private final HashMap<String, String> MAP =<fold text=' ::' expand='true'> </fold>new<fold text='' expand='true'> </fold><fold text='' expand='true'>HashMap<></fold><fold text='' expand='true'>()</fold>;
+        private final HashMap<String, String> MAP2 =<fold text=' ::' expand='true'> </fold>new<fold text='' expand='true'> </fold><fold text='' expand='true'>HashMap<fold text='<~>' expand='false'><String, String></fold></fold><fold text='' expand='true'>()</fold>;
+        private final Map<String, String> MAP3 = new HashMap<>();
+
+        private final List<String> LIST = new ArrayList<>();
+        private final List<String> LIST2 = List.of("1");
+
+        public final ConstClass SELF_PARAM_1 =<fold text=' ::' expand='true'> </fold>new<fold text='' expand='true'> </fold><fold text='' expand='true'>ConstClass</fold>(true);
+        public final ConstClass SELF_PARAM_2 =<fold text=' ::' expand='true'> </fold>new<fold text='' expand='true'> </fold><fold text='' expand='true'>ConstClass</fold>(false, LIST2.get(0));
+
+        public final ConstClass SELF_SUBCLASS_MORE_FIELD = new ConstClass() <fold text='{...}' expand='true'>{
+            int i = 1;
+        }</fold>;
+        public final ConstClass SELF_SUBCLASS_MORE_FUNC = new ConstClass() <fold text='{...}' expand='true'>{
+            public void setOk(boolean ok) <fold text='{}' expand='true'>{
+            }</fold>
+        }</fold>;
+
+        protected ConstClass SELF_NULL = null;
+        protected ConstClass EMPTY;
+    }</fold>
+
+    static class ConstClass <fold text='{...}' expand='true'>{
+        boolean ok;
+        String string;
+
+        public ConstClass() <fold text='{}' expand='true'>{
+        }</fold>
+
+        public ConstClass(boolean ok)<fold text=' { ' expand='false'> {
+            </fold>this.ok = ok;<fold text=' }' expand='false'>
+        }</fold>
+
+        public ConstClass(boolean ok, String string) <fold text='{...}' expand='true'>{
+            this.ok = ok;
+            this.string = string;
+        }</fold>
+
+        @Override
+        public String toString() <fold text='{...}' expand='true'>{
+            return new StringJoiner(", ", ConstClass.class.getSimpleName() + "[", "]")
+                    .add("string='" + string + "'")
+                    .toString();
+        }</fold>
+    }</fold>
+
+    static class SubConstClass extends ConstClass <fold text='{...}' expand='true'>{
+    }</fold>
+
+    static final class SubConstClass2 extends ConstClass <fold text='{...}' expand='true'>{
+    }</fold>
+}